refactor(dependency-injection): migrate from invectify to inversify
Update all imports and configuration files to use InversifyJS instead of the custom invectify implementation. This provides better type safety and is a more standard solution for dependency injection in TypeScript projects.
