+++ services/elastic-search-service/src/Inversify/types.ts
